Actualizar listado de salidas
